refactor(pymeet-rest): moved to corresponding package
