+++ pymeet-rest/src/pymeet/domain/errors.py
